+++ includes/Achievement.php
@@ -2,10 +2,13 @@
 
 namespace MediaWiki\Extension\AchievementBadges;
 
+use BetaFeatures;
 use EchoEvent;
 use ManualLogEntry;
+use MediaWiki\MediaWikiServices;
 use MWException;
 use SpecialPage;
+use User;
 
 class Achievement {
 	/**
@@ -24,7 +27,7 @@ class Achievement {
 		}
 
 		$user = $info['user'];
-		if ( !Utils::isAchievementBadgesAvailable( $user ) ) {
+		if ( !self::isAchievementBadgesAvailable( $user ) ) {
 			return;
 		}
 		$key = $info['key'];
@@ -67,5 +70,28 @@ class Achievement {
 			],
 			'agent' => $user,
 		] );
+	}
+
+	/**
+	 * @param User $user
+	 * @return bool
+	 */
+	public static function isAchievementBadgesAvailable( User $user ) {
+		$config = MediaWikiServices::getInstance()->getMainConfig();
+
+		$configEnabled = $config->get( Constants::CONFIG_KEY_ACHIEVEMENT_BADGES_ENABLE_BETA_FEATURE );
+		$userOptionEnabled = $configEnabled &&
+			BetaFeatures::isFeatureEnabled( $user, Constants::PREF_KEY_ACHIEVEMENT_ENABLE );
+
+		if ( !$configEnabled ) {
+			// If AchievementBadges is not a beta feature, it is available to everyone.
+			return true;
+		}
+		if ( $user->isRegistered() && $userOptionEnabled ) {
+			// If AchievementBadges is a beta feature, only a registered user which enables the feature
+			// has access the feature.
+			return true;
+		}
+		return false;
 	}
 }