+++ includes/AchievementRechecker.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace MediaWiki\Extension\AchievementBadges;
+
+use User;
+
+class AchievementRechecker {
+	/**
+	 * @param User $user
+	 * @return bool
+	 */
+	public static function checkAlwaysTrue( User $user ) {
+		return true;
+	}
+}